requirements.txt
alembic==0.9.9
aniso8601==3.0.2
atomicwrites==1.1.5
attrs==18.1.0
bleach==1.5.0
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
enum34==1.1.6
eventlet==0.23.0
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
graphene==2.1.2
graphene-sqlalchemy==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
greenlet==0.4.13
html5lib==0.9999999
idna==2.7
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
more-itertools==4.2.0
numpy==1.13.3
pluggy==0.6.0
promise==2.1
protobuf==3.5.0.post1
psycopg2==2.7.5
py==1.5.3
pytest==3.6.2
python-crontab==2.3.4
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.19.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.8
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
typing==3.6.4
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
